requirements_powerup.txt
# HyperbolicLearner Transcendent System - Power-Up Requirements
# Install with: pip install -r requirements_powerup.txt

# Core AI & ML (Already included in base)
torch>=2.1.0
torchvision>=0.16.0
torchaudio>=2.1.0
transformers[torch]>=4.35.0
diffusers>=0.24.0
accelerate>=0.24.0
optimum>=1.14.0

# Advanced Computer Vision
opencv-python>=4.8.0
pillow>=10.0.0
scikit-image>=0.22.0
mediapipe>=0.10.0
ultralytics>=8.0.0
segment-anything>=1.0
clip-by-openai>=1.0

# Natural Language Processing
spacy>=3.7.0
nltk>=3.8.0
sentence-transformers>=2.2.0
polyglot>=16.07.04
googletrans>=3.1.0
langdetect>=1.0.9
textblob>=0.17.1

# Audio Processing
librosa>=0.10.0
soundfile>=0.12.0
pyaudio>=0.2.11
speech-recognition>=3.10.0
pyttsx3>=2.90
whisper-openai>=20230918
musicgen>=1.0.0

# Web Automation & Scraping
selenium>=4.15.0
playwright>=1.40.0
requests>=2.31.0
beautifulsoup4>=4.12.0
scrapy>=2.11.0
httpx>=0.25.0

# Desktop Automation
pyautogui>=0.9.54
pygetwindow>=0.0.9
keyboard>=0.13.5
mouse>=0.7.1
pynput>=1.7.6
accessibility>=0.1.0

# Mobile Automation
appium-python-client>=3.1.0
android-automation>=1.0.0
ios-automation>=1.0.0

# Database & Storage
asyncpg>=0.29.0
redis>=5.0.0
elasticsearch>=8.11.0
neo4j>=5.14.0
mongodb>=4.5.0
sqlalchemy>=2.0.0
alembic>=1.12.0

# Cloud Infrastructure
boto3>=1.34.0
azure-sdk>=1.0.0
google-cloud-storage>=2.10.0
kubernetes>=28.1.0
docker>=6.1.0
terraform-python>=1.0.0

# API & Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0
pydantic[email]>=2.5.0
starlette>=0.27.0
httpx>=0.25.0

# Monitoring & Analytics
prometheus-client>=0.19.0
grafana-client>=3.7.0
datadog>=0.48.0
sentry-sdk>=1.38.0
structlog>=23.2.0
loguru>=0.7.2

# Security & Encryption
cryptography>=41.0.0
pyjwt>=2.8.0
oauth2lib>=1.0.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0

# Blockchain & Web3
web3>=6.11.0
eth-account>=0.9.0
solana-py>=0.30.0
bitcoin>=1.1.42
ethereum>=2.3.2

# IoT & Hardware
paho-mqtt>=1.6.0
pyserial>=3.5
RPi.GPIO>=0.7.1
adafruit-circuitpython>=8.2.0
bleak>=0.21.0

# Scientific Computing
numpy>=1.25.0
scipy>=1.11.0
pandas>=2.1.0
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.17.0
jupyter>=1.0.0

# Machine Learning Extensions
xgboost>=2.0.0
lightgbm>=4.1.0
catboost>=1.2.0
optuna>=3.4.0
hyperopt>=0.2.7
ray[tune]>=2.8.0

# Financial & Trading
yfinance>=0.2.18
alpha-vantage>=2.3.1
ccxt>=4.1.0
ta-lib>=0.4.28
pandas-ta>=0.3.14b

# Document Processing
PyPDF2>=3.0.0
python-docx>=1.1.0
openpyxl>=3.1.0
python-pptx>=0.6.22
markdown>=3.5.0
pypandoc>=1.12

# Image & Video Processing
moviepy>=1.0.3
imageio>=2.31.0
ffmpeg-python>=0.2.0
wand>=0.6.13

# Development & Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.9.0
isort>=5.12.0
mypy>=1.6.0
pre-commit>=3.5.0

# Performance & Optimization
cython>=3.0.0
numba>=0.58.0
dask>=2023.10.0
joblib>=1.3.0

# Networking & Communication
aiohttp>=3.9.0
websocket-client>=1.6.0
paramiko>=3.3.0
fabric>=3.2.0

# Utility & Helpers
click>=8.1.0
rich>=13.6.0
tqdm>=4.66.0
python-dotenv>=1.0.0
pyyaml>=6.0.1
toml>=0.10.2
configparser>=6.0.0
